Propose fix for tallying of votes
Unless I made a mistake in following the instructions, I had trouble getting an actual tally of total votes to be calculated. I rewrote the functions used to store votes in the manner proposed in this change.

It does calculate the votes properly. However, since it doesn't use the socket information, I doubt that the disconnect function still works.